+++ file-sharing/src/tabs/iSCSI/ui/screens/portal/PortalTable.vue
@@ -13,9 +13,9 @@
           @close-editor="
             () => {
               showAddPortal = false;
-              actions.refreshPortals();
             }
           "
+          @created='actions.refreshPortals()'
         />
       </div>
     </div>
@@ -80,5 +80,5 @@ const refreshPortals = () => {
 
 const actions = wrapActions({ refreshPortals: refreshPortals });
 
-actions.refreshPortals();
+//actions.refreshPortals();
 </script>